remove email added

+++ app/Console/Commands/RemoveEmail.php
@@ -45,22 +45,27 @@ class RemoveEmail extends Command
         ini_set('memory_limit', -1);
         ini_set('mysql.connect_timeout', 600);
         ini_set('default_socket_timeout', 600);
-        $available_email = Emails::all(['email']);
-        $email_array = $available_email->pluck('email');
-        $plucked_all = $email_array->all();
-        $emails = AvailableEmail::whereIn('email',$plucked_all)->take(10)->get();
-        foreach($emails AS $email){
-            $email_db = Emails::where('email',$email->email)->get();
-            $matched_contact_id = $email_db->first()->matched_contact_id;
-            echo $matched_contact_id.",";
-            $matched_contact = MatchedContact::where('id',$matched_contact_id)->get();
-            $mt = $matched_contact->first();
-            $mt->email = $email_db->first()->email;
-            $mt->email_status = 'valid';
-            $mt->email_validation_date = '2018-09-01 00:00:00';
-            $saved = $mt->save();
-            if($saved){
-               Emails::where('matched_contact_id',$matched_contact_id)->delete(); 
+        $emails = Emails::all(['email']);
+        $emails_array = $emails->pluck('email');
+        $emails_all = $emails_array->all();
+        $available_email = AvailableEmail::all(['email']);
+        $available_array = $available_email->pluck('email');
+        $available_all = $available_array->all();
+        $result = array_intersect($emails_all,$available_all);
+        if(count($result) > 0){
+            foreach($result AS $email){
+                $email_db = Emails::where('email',$email)->get();
+                $matched_contact_id = $email_db->first()->matched_contact_id;
+                echo $matched_contact_id.",";
+                $matched_contact = MatchedContact::where('id',$matched_contact_id)->get();
+                $mt = $matched_contact->first();
+                $mt->email = $email_db->first()->email;
+                $mt->email_status = 'valid';
+                $mt->email_validation_date = '2018-09-01 00:00:00';
+                $saved = $mt->save();
+                if($saved){
+                   Emails::where('matched_contact_id',$matched_contact_id)->delete(); 
+                }
             }
         }
         UtilDebug::debug("End processing");
